requirements.txt
requests
pymongo
openai
beautifulsoup4
lxml
python-dotenv
